+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/BuildTest.cs
@@ -4212,27 +4212,5 @@ namespace UnnamedProject
 				}
 			}
 		}
-
-		[Test]
-		public void InstallAndroidDependenciesTest ()
-		{
-			if (!CommercialBuildAvailable)
-				Assert.Ignore ("Not required on Open Source Builds");
-			var old = Environment.GetEnvironmentVariable ("ANDROID_SDK_PATH");
-			try {
-				string sdkPath = Path.Combine (Root, "temp", TestName, "android-sdk");
-				Environment.SetEnvironmentVariable ("ANDROID_SDK_PATH", sdkPath);
-				var proj = new XamarinAndroidApplicationProject ();
-				using (var b = CreateApkBuilder ()) {
-					string defaultTarget = b.Target;
-					b.Target = "InstallAndroidDependencies";
-					Assert.IsTrue (b.Build (proj, parameters: new string [] { "AcceptAndroidSDKLicenses=true" }), "InstallAndroidDependencies should have succeeded.");
-					b.Target = defaultTarget;
-					Assert.IsTrue (b.Build (proj), "build should have succeeded.");
-				}
-			} finally {
-				Environment.SetEnvironmentVariable ("ANDROID_SDK_PATH", old);
-			}
-		}
 	}
 }
