+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Xamarin.Android.Build.Tests.Shared.projitems
@@ -32,5 +32,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\EnvironmentHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\PackagingUtilsTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)WearTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)AndroidDependenciesTests.cs" />
   </ItemGroup>
 </Project>
